Refactors ui : moves to proper package
